Fix to only update the cell data if the cell actually ran
If a cell did not run, data like output_tags for that cell may not be correct. This happens when multiple cell executions happen at once in an order we do not control (i.e. runSelected) because the last cell may do nothing and return the original cell data, wiping out the updated cell data from a previous execution.